clean-#1-domains 구축
