+++ domains/repositories.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+from .objects import Story
+
+class StoryReader(ABC):
+    @abstractmethod
+    def read(self, flat_form, category ) -> Story:
+        raise NotImplementedError()
