+++ news/views/story_views.py
@@ -11,14 +11,16 @@ class StoryListView(ListAPIView):
     queryset = Story.objects.all()
     serializer_class = StoryListSZ
 
-    def get_queryset(self):
-        flat_form_title = self.kwargs.get('flatform')
-        category_title = self.kwargs.get('category')
-
-        if not FlatForm.objects.filter(title=flat_form_title):
-            raise exceptions.NotFound(detail='flatform not exists')
-
-        if not Category.objects.filter(title=category_title):
-            raise exceptions.NotFound(detail='category not exists')
-
-        return self.queryset.filter(flat_form__title=flat_form_title, category__title=category_title)
+    def list(self, request, *args, **kwargs):
+        pass
+    # def get_queryset(self):
+    #     flat_form_title = self.kwargs.get('flatform')
+    #     category_title = self.kwargs.get('category')
+    #
+    #     if not FlatForm.objects.filter(title=flat_form_title):
+    #         raise exceptions.NotFound(detail='flatform not exists')
+    #
+    #     if not Category.objects.filter(title=category_title):
+    #         raise exceptions.NotFound(detail='category not exists')
+    #
+    #     return self.queryset.filter(flat_form__title=flat_form_title, category__title=category_title)
